updated py imageSearch to use casefold
Realized I wasn't actually searching the casefolded version of the images list. Changing that then broke finding the image index to remove the image because the cases weren't matching. So I realized the before the delete step, the indexes in both arrays would match so I could use the index I'm currently on to append the uncasefolded version to the list of things I wanted to remove. This way the next step could actually match the images.

I believe the test wouldn't accept the output if the cases didn't match what was put in. So I couldn't just return the casefolded version and had to make that adjustment.